# backend/swagger.yaml
openapi: 3.0.0
paths:
...
  '/users/{id}':
    get:
      operationId: UsersController_findById
      summary: ''
      description: Return all properties for one user OR throw an error
      parameters:
        - name: id
          required: true
          in: path
          description: Id of the user we want to find in the database.
          allowEmptyValue: false
          examples:
            a:
              summary: User id is 1
              value: '1'
            b:
              summary: User id is 42
              value: '42'
          schema:
            type: string
        - name: friend
          required: true
          in: query
          description: Get the user friend ?
          allowEmptyValue: true
          examples:
            a:
              summary: '?friend=true'
              value: 'true'
            b:
              summary: '?friend=false'
              value: 'false'
          schema:
            type: string
        - name: channel
          required: true
          in: query
          description: Get the user channel ?
          allowEmptyValue: true
          examples:
            a:
              summary: '?channel=true'
              value: 'true'
            b:
              summary: '?channel=false'
              value: 'false'
          schema:
            type: string
      responses:
        '200':
          description: 'Return an object containing the user info + ?{friend/channel}'
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - username
                  - AvatarUrl
                properties:
                  id:
                    type: number
                  username:
                    type: string
                  avatarUrl:
                    type: string
                  friends:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                        username:
                          type: string
                        avatarUrl:
                          type: string
                  channelsProfiles:
                    type: array
                    items:
                      type: object
                      properties:
                        channelId:
                          type: number
                        role:
                          type: string
                        channel:
                          type: object
                          items:
                            type: object
                            properties:
                              channelName:
                                type: string
      tags:
        - Users
  '/users/{id}/friends':
    get:
      operationId: UsersController_allFriends
      summary: ''
      description: Return an array of friends
      parameters:
        - name: id
          required: true
          in: path
          description: Id of the user for whom we want to get the list of his friend users.
          allowEmptyValue: false
          examples:
            a:
              summary: User id is 1
              value: '1'
            b:
              summary: User id is 42
              value: '42'
          schema:
            type: string
      responses:
        '200':
          description: A FAIRE
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - username
                  - AvatarUrl
                properties:
                  id:
                    type: number
                  username:
                    type: string
                  AvatarUrl:
                    type: string
        '400':
          description: 'This description is for a 400 response. It is returned when additional query params are passed, or when the <b>useExclamation</b>-Argument could not be parsed as a boolean.'
        '417':
          description: 'One can also provided a Status-Code directly, as seen here'
      tags:
        - Users
...
  /users/addFriend:
    post:
      operationId: UsersController_addFriend
      summary: ''
      description: Add new friend
      parameters: []
      requestBody:
        required: true
        description: ''
        content:
          application/json:
            schema:
              type: string
            examples:
              a:
                summary: Invalid Body
                description: '{ friendId: id } required. (id: number)'
                value: {}
              b:
                summary: Valid Body
                value:
                  friendId: 1
      responses:
        '200':
          description: A FAIRE
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - username
                  - AvatarUrl
                properties:
                  id:
                    type: number
                  username:
                    type: string
                  AvatarUrl:
                    type: string
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntity'
        '400':
          description: 'This description is for a 400 response. It is returned when additional query params are passed, or when the <b>useExclamation</b>-Argument could not be parsed as a boolean.'
      tags:
        - Users
  /users/addBlockedUser:
    post:
      operationId: UsersController_addBlockedUser
      summary: ''
      description: Add new blocked user
      parameters: []
      requestBody:
        required: true
        description: ''
        content:
          application/json:
            schema:
              type: string
            examples:
              a:
                summary: Invalid Body
                value: {}
              b:
                summary: Valid Body
                value:
                  blockedId: 1
      responses:
        '200':
          description: A FAIRE
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - username
                  - AvatarUrl
                properties:
                  id:
                    type: number
                  username:
                    type: string
                  AvatarUrl:
                    type: string
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntity'
        '400':
          description: 'This description is for a 400 response. It is returned when additional query params are passed, or when the <b>useExclamation</b>-Argument could not be parsed as a boolean.'
      tags:
        - Users
  '/users/deleteFriend/{id}':
    delete:
      operationId: UsersController_deleteFriend
      summary: ''
      description: Delete one friend
      parameters:
        - name: id
          required: true
          in: path
          description: Id of the user for whom we want to delete a friend
          allowEmptyValue: false
          examples:
            a:
              summary: User id is 1
              value: '1'
            b:
              summary: User id is 42
              value: '42'
          schema:
            type: string
        - name: friendId
          required: true
          in: query
          description: Id of this friend
          allowEmptyValue: false
          examples:
            a:
              summary: Friend id is 2
              value: '2'
            b:
              summary: Friend id is 42
              value: '42'
          schema:
            type: string
      responses:
        '200':
          description: SUCCESS response
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - username
                  - AvatarUrl
                properties:
                  id:
                    type: number
                  updatedAt:
                    type: string
                  username:
                    type: string
                  AvatarUrl:
                    type: string
        '400':
          description: 'This description is for a 400 response. It is returned when additional query params are passed, or when the <b>useExclamation</b>-Argument could not be parsed as a boolean.'
      tags:
        - Users
  '/users/deleteBlockedUser/{id}':
    delete:
      operationId: UsersController_deleteBlockedUser
      summary: ''
      description: Delete one blocked user
      parameters:
        - name: friendId
          required: true
          in: query
          schema:
            type: string
        - name: id
          required: true
          in: path
          description: Id of the user for whom we want to get the list of his friend users.
          allowEmptyValue: false
          examples:
            a:
              summary: User id is 1
              value: '1'
            b:
              summary: User id is 42
              value: '42'
          schema: {}
      responses:
        '200':
          description: A FAIRE
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - username
                  - AvatarUrl
                properties:
                  id:
                    type: number
                  username:
                    type: string
                  AvatarUrl:
                    type: string
        '400':
          description: 'This description is for a 400 response. It is returned when additional query params are passed, or when the <b>useExclamation</b>-Argument could not be parsed as a boolean.'
      tags:
        - Users
...
  '/users/{id}/games':
    get:
      operationId: UsersController_getGameOfUser
      summary: Get games of user with is id
      parameters:
        - name: id
          required: true
          in: path
          description: Id of the player
          allowEmptyValue: false
          examples:
            a:
              summary: Palyer with id 1
              value: '1'
            b:
              summary: Player with id 27
              value: '27'
          schema:
            type: number
        - name: limit
          required: true
          in: query
          description: Limit of games to get (default and maximum 50)
          allowEmptyValue: true
          examples:
            a:
              summary: Get 37 games
              value: '37'
          schema:
            type: number
        - name: idFirst
          required: true
          in: query
          description: Id to start the next get
          allowEmptyValue: true
          examples:
            a:
              summary: Start after game with id 27
              value: '27'
          schema:
            type: number
        - name: asWin
          required: true
          in: query
          description: Only get games where the user win or lose
          allowEmptyValue: true
          examples:
            a:
              summary: Game win
              value: 'true'
            b:
              summary: Game Lose
              value: 'false'
          schema:
            type: boolean
      responses:
        '200':
          description: 'Return an array containing {limit} games of one player, from idFirst from the youngest game to the oldest game'
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - winnerId
                    - isFinish
                    - createdAt
                    - players
                  properties:
                    id:
                      type: number
                    winnerId:
                      type: number
                    isFinish:
                      type: boolean
                    createdAt:
                      type: string
                      format: date-time
                    players:
                      type: array
                      items:
                        type: object
                        required:
                          - id
                          - gameId
                          - userId
                          - asWin
                          - score
                        properties:
                          id:
                            type: number
                          gameId:
                            type: number
                          userId:
                            type: number
                          asWin:
                            type: boolean
                          score:
                            type: number
        '400':
          description: '400 is returned when : <br/>- Additional query params are passed <br/>- Argument could not be parsed as number <br/>- ''idFirst'' or ''limit'' is negative <br/>- ''asWin'' is not ''true'' or ''false'''
      tags:
        - Users
...
info:
  title: TRANSCENDENCE
  description: Our 42 project ft_transcendence API description.
  version: v1
  contact: {}
...
servers: []
components:
  securitySchemes:
    TransandanceToken:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Enter access token for transandance
  schemas:
    UserEntity:
      type: object
      properties:
        id:
          type: number
        username:
          type: string
        avatarUrl:
          type: string
        updatedAt:
          format: date-time
          type: string
        friends:
          type: array
          items:
            type: object
        channelsProfiles:
          type: array
          items:
            type: object
        createdAt:
          format: date-time
          type: string
        id42:
          type: number
      required:
        - id
        - username
        - avatarUrl
        - updatedAt
        - createdAt
        - id42
...